backend/requirements.txt
torch==1.9.0
torchvision==0.10.0
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
scikit-learn==0.24.2
tensorboard==2.5.0
Flask==1.1.2
werkzeug==0.16.1
jinja2==2.11.3
markupsafe==2.0.1
itsdangerous==1.1.0
protobuf==3.20.0
flask-cors==4.0.1
python-dotenv==0.19.0